Semi-working version.  Throws no errors
It decides whether to pull the html locally or from the server.

And it can query the db for addOn attorneyIds but it isn't saving the addOn id html files.  And it is creating duplicate NAC with different current attorneyIDs in the NACs array; even when the NACs aren't' duplicated on the clerk schedule pages.