I have gotten a lot further wit the code.  Creating the schedule class scrapes the clerk's site if the db date is out of date.  It inserts the NACs into the db.  But I have not written the code to select the NACs from the db. That is the next step.
Consider relying on the table that maps user bar numbers to case numbers to output NACS.  But think hard about this special case: A attorney of record is on a case with co-counsel that is an add on bar number. If the user is taken off the case the table should still link his barnumber with the case number because of the add on.  Maybe a reference counter is needed.